refactor classifier
